Move tooltip code from field to bas widget class.
